% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytest_utils.R
\name{activate_session}
\alias{activate_session}
\title{Activate an existing session within the test mode of MOATR}
\usage{
activate_session(
  app,
  session_name,
  session_dirname = ".moatr_files_test",
  module_id = "first"
)
}
\arguments{
\item{app}{A `ShinyDriver` object}

\item{session_name}{String with the name of the session to activate.}

\item{session_dirname}{String for directory name of MOATR session files.
Defaults to `.moatr_files_test` since that is the custom directory
used for testing the app in `shinytest`.}

\item{module_id}{Namespace ID of the welcome module. Default is `first`}
}
\value{
invisible TRUE
}
\description{
Activate an existing session within the test mode of MOATR
}
