% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytest_utils.R
\name{configure_app}
\alias{configure_app}
\title{Customize settings for shinytest app object}
\usage{
configure_app(app, test_name, test_files = NULL, verbose = FALSE)
}
\arguments{
\item{test_name}{String for the name of the test being performed}

\item{A}{`ShinyDriver` object}
}
\value{
list with the paths to the snapshot, screenshot, and log directories
}
\description{
Customize settings for shinytest app object
}
