% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytest_utils.R
\name{deactivate_session}
\alias{deactivate_session}
\title{Deactivate the session currently active in MOATR}
\usage{
deactivate_session(
  app,
  session_dirname = ".moatr_files_test",
  module_id = "first"
)
}
\arguments{
\item{app}{A `ShinyDriver` object}

\item{session_dirname}{String for directory name of MOATR session files.
Defaults to `.moatr_files_test` since that is the custom directory
used for testing the app in `shinytest`.}

\item{module_id}{Namespace ID of the welcome module. Default is `first`}
}
\description{
Deactivate the session currently active in MOATR
}
