% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytest_utils.R
\name{dismiss_alert}
\alias{dismiss_alert}
\title{Automatically dismiss an alert/modal dialog in shiny app}
\usage{
dismiss_alert(app, type, keypress = "escape")
}
\arguments{
\item{app}{A `ShinyDriver` object}

\item{type}{string indicating the type of alert to dismiss. Currently the
following types are supported:
  * `modal`: Modal dialog that is typically produced by `shiny::showModal` and there
             is not a custom input button to press other than "Dismiss"
  * `sweetalert`: Alert produced by the `shinyalert` package and there is not a 
                  custom input button to press other than "Ok" or something similar.}

\item{keypress}{which key to simulate pressing found in the `webdriver::key` list.
Default is `escape` for the escape key.}
}
\value{
invisible TRUE
}
\description{
This function is meant to be used within a `shinytest` script when 
a popup alert or modal appears on the screen and it has the capability
of being dismissed via a keypress such as the escape key
}
