% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytest_utils.R
\name{log_message}
\alias{log_message}
\title{Print diagnostic log messages to test log file}
\usage{
log_message(step_name, session_settings, log_lvl = "INFO", ...)
}
\arguments{
\item{step_name}{String for name of step (typically formatted like `step1`)}

\item{session_settings}{list with customized test settings imported from `session_config.yml`}

\item{log_lvl}{Log level used by `loggit` default is `INFO`}

\item{...}{Additional variables (scalars) to pass to `loggit` function}
}
\value{
Nothing
}
\description{
Print diagnostic log messages to test log file
}
