% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytest_utils.R
\name{module}
\alias{module}
\title{Load a linux module into the R session}
\usage{
module(...)
}
\arguments{
\item{...}{typically the string portion of a `module` command. For example,
if the command to load the R module is `module load R-qualified`, then the 
string that should be supplied is `"load R-qualified"`.}
}
\value{
nothing
}
\description{
Load a linux module into the R session
}
\examples{
\dontrun{
module("--ignore_cache load phantomjs")
}
}
