% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytest_utils.R
\name{prepare_testdir}
\alias{prepare_testdir}
\title{Prepare a custom MOATR session directory appropriate for automated testing}
\usage{
prepare_testdir(
  testname,
  user = Sys.getenv("USER"),
  moatr_version = "2",
  session_location = "/lrlhps/data/moatr_files",
  session_config = "session-config.yml",
  set_env = TRUE
)
}
\arguments{
\item{testname}{name of the test used in the session. Specific settings for each of these
are contained in the `session-config.yml` file. Choices are the following:
* `test_welcome`
* `test_design`
* `test_simulate_upload`
* `test_simulate_schemeconfig`
* `test_simulate_hpc`
* `test_postprocess_summary`
* `test_postprocess_combo`
* `test_postprocess_value`}

\item{user}{User ID}

\item{moatr_version}{string with MOATR version number}

\item{session_location}{Directory path to where the pre-made files that are used
as setup files for each test are given.}

\item{session_config:}{Name of session configuration file to import with `config` package.
Default is `session-config.yml`.}
}
\value{
invisible TRUE
}
\description{
This is a convenience function to initialize a MOATR session directory 
to `/lrlhps/users/<user_id>/.moatr_files_test` based on pre-made session
directories stored in the MOATR isilon location.
}
