% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytest_utils.R
\name{setup_hpc}
\alias{setup_hpc}
\title{Set up system environment for HPC simulation launch}
\usage{
setup_hpc()
}
\value{
`TRUE` invisibly
}
\description{
This simple function loads the `cluster-setup` module so HPC simulates performed
on a compute node (typically within`shinytest` execution) can perform correctly
}
