% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytest_utils.R
\name{setup_shinytest}
\alias{setup_shinytest}
\title{Set up system environment for `shinytest`}
\usage{
setup_shinytest()
}
\value{
`TRUE` invisibly
}
\description{
This simple function loads the `phantomjs` module so `shinytest` can 
perform as expected
}
