% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinytest_utils.R
\name{update_rdata}
\alias{update_rdata}
\title{Update session directory and HPC workdir paths in pre-work test session directories}
\usage{
update_rdata(
  testname,
  version_number = "2",
  session_id = "testsession1",
  session_config_file = "session-config.yml"
)
}
\arguments{
\item{testname}{String with name of test session directory}

\item{version_number}{MOATR version number}

\item{session_id}{String with the pre-saved session ID that is used in the test}

\item{session_config_file}{YAML config file with settings for each test name}
}
\value{
Nothing
}
\description{
This function is necessary because if you use one of the test scripts interactive to
generate session files and copy those to the test session source directory, it will
have the session directory and HPC workdir path from that interactive session. When
using those files in a "new" test that builds upon those files, there will be a mismatch
in those directory paths. This function simply updates the appropriate slots in the 
`r_data` list object with the correct paths.
}
\examples{
library(purrr)
test_names <- c("test_design", "test_demo", "test_simulate_upload",
"test_simulate_schemeconfig", "test_simulate_hpc", "test_postprocess_summary",
"test_postprocess_combo", "test_postprocess_value")

walk(test_names, ~update_rdata(.x))
}
